requirements.txt
streamlit
pypdf
reportlab
Pillow
python-dotenv
pymupdf